requirements.txt
SQLAlchemy==2.0.27
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
python-telegram-bot==20.8
Telethon==1.34.0
python-docx==0.8.11
PyPDF2==3.0.1
spacy==3.7.2
phonenumbers==8.13.30
python-magic-bin==0.4.14
natasha==1.6.0
gpt4all==2.0.2
yargy==0.16.0
razdel==0.5.0
transformers==4.36.2
torch==2.1.2
pdfplumber==0.10.3 